requirements.txt
asyncio==3.4.3
discord.py==1.3.3
mcstatus===3.1.0
python-dotenv==0.13.0